backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
celery==5.3.6
redis==5.0.4
anthropic==0.28.0
boto3==1.34.87
supabase==2.4.5
httpx==0.27.0
pandas==2.2.2
aiofiles==23.2.1
python-dotenv==1.0.1
